Restructured project to fix execution
