+++ cmd/mosh.go
@@ -13,13 +13,13 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package main
+package cmd
 
 import (
 	"fmt"
 	"time"
 
-	"optimup/jpegrej/pkg"
+	"github.com/optimup/jpegrej/pkg"
 
 	"github.com/spf13/cobra"
 )
