+++ main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/optimup/jpegrej/cmd"
+
+func main() {
+    cmd.Execute()
+}
